Made code more similar to the manual process we use for configuring a control plane
